requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-jose
passlib[bcrypt]
slowapi
groq